Update README.md
Welcome to Linux Basics for Beginners, a comprehensive guide to help new users navigate the Linux command-line environment. This repository is perfect for anyone starting their Linux journey and aiming to build a strong foundation. Identifying the right commands for specific tasks can be challenging at first, but these guidelines are designed to simplify the process and accelerate learning for beginners.